requirements.txt
pytesseract
pyautogui
pillow
